fixed bug for the binary tree height file

+++ 9-binary_tree_height.c
@@ -8,8 +8,8 @@
 {
     if (tree == NULL)
         return (0);
-    int leftHeight = binary_tree_height(tree->left);
-    int rightHeight = binary_tree_height(tree->right);
+    size_t leftHeight = binary_tree_height(tree->left);
+    size_t rightHeight = binary_tree_height(tree->right);
     if (leftHeight >= rightHeight)
         return (leftHeight + 1);
     else
